+++ packages/frontend/src/components/layout/layout.component.tsx
@@ -5,6 +5,7 @@ import { Navigate } from 'react-router-dom';
 import { ROUTES } from '@constants/routes';
 import { useMeRequest } from '@api/index';
 import { Toast } from '../toast/toast.component';
+import { SplashScreen } from '@pages/splash/splash.page';
 
 /**
  * @description Layout HOC for base settings for all pages
@@ -22,7 +23,7 @@ export const Layout: FC<PropsWithChildren<ILayout>> = ({
 }) => {
   const { isLoading, error, data } = useMeRequest();
 
-  if (isLoading && !error && !data) return <h1>Loading...</h1>;
+  if (isLoading && !error && !data) return <SplashScreen />;
   if (
     (error?.name === 'UNAUTHORIZED' && protectedFrom === 'anonymous') ||
     (data && protectedFrom === 'authorized')
