+++ packages/frontend/src/components/ui-kit/loader/loader.module.css
@@ -1,45 +1,45 @@
-.loader{
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 6px;
-    width: 100%;
+.loader {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 6px;
+  width: 100%;
 }
 
-.loader__gradient--default{
-    background: linear-gradient(to right, #FF6B00, #ECF100);
+.loader__gradient--default {
+  background: linear-gradient(to right, #ff6b00, #ecf100);
 }
 
-.loader__gradient--blueToСyan{
-    background: linear-gradient(to right, #006ACC, #00C2FF);
+.loader__gradient--blueToСyan {
+  background: linear-gradient(to right, #006acc, #00c2ff);
 }
 
-.loader__gradient--violetToMagenta{
-    background: linear-gradient(to right, #6B00D7, #C90085);
+.loader__gradient--violetToMagenta {
+  background: linear-gradient(to right, #6b00d7, #c90085);
 }
 
-.loader__hidden{
-    position: absolute;
-    width: 100%;
-    left: 0;
-    background-color: #FAFAFA;
-    z-index: 2;
-    animation: 1.5s cubic-bezier(0.19, 0.5, 1, 0.7) hidden forwards;
+.loader__hidden {
+  position: absolute;
+  width: 100%;
+  left: 0;
+  background-color: #fafafa;
+  z-index: 2;
+  animation: 10s cubic-bezier(0.01, 1.3, 1, 0.9) hidden forwards;
 }
 
 @keyframes hidden {
-    0%{
-        position: absolute;
-        width: 100%;
-        left: 0;
-        background-color: #FAFAFA;
-        z-index: 2;
-    }
-    100%{
-        position: absolute;
-        width: 0;
-        left: 100vw;
-        background-color: #FAFAFA;
-        z-index: 2;
-    }
+  0% {
+    position: absolute;
+    width: 100%;
+    left: 0;
+    background-color: #fafafa;
+    z-index: 2;
+  }
+  100% {
+    position: absolute;
+    width: 0;
+    left: 100vw;
+    background-color: #fafafa;
+    z-index: 2;
+  }
 }
