+++ packages/frontend/src/components/ui-kit/spinner/spinner.component.tsx
@@ -4,7 +4,7 @@ import css from './spinner.module.css';
 import { ISpinner } from './spinner';
 import { clx } from '../../../utils/clx';
 
-export const SpinnerComponent: FC<ISpinner> = ({ theme = 'default' }) => {
+export const Spinner: FC<ISpinner> = ({ theme = 'default' }) => {
   const ldsRollerTheme = clx(css[`lds__roller`], css[`lds__roller--${theme}`]);
 
   return (
