+++ packages/frontend/src/pages/error/error-info/error-info-status/error-info-status.component.tsx
@@ -1,13 +1,12 @@
-import {FC} from 'react';
-import css from './error-info-status.module.css'
+import { FC } from 'react';
+import css from './error-info-status.module.css';
 import { IErrorInfoStatus } from '@pages/error/error-info/error-info-status/error-info-status';
-export const ErrorInfoStatusComponent: FC<IErrorInfoStatus> = ({status}) => {
 
+export const ErrorInfoStatus: FC<IErrorInfoStatus> = ({ status }) => {
   return (
     <div>
       <p className={css.errorInfo__errorStatus}>{`Ошибка ${status}`}</p>
       <p className={css.errorInfo__explanation}>Что-то пошло не так</p>
     </div>
-
   );
 };
