+++ packages/frontend/src/pages/error/error-info/error-info-title/error-info-title.module.css
@@ -1,8 +1,4 @@
-.errorInfo__title{
-    font-size: 1.8rem;
-    text-align: center;
-}
-
-.errorInfo__link{
-    margin-left: 5px;
+.errorInfo__title {
+  font-size: 1.8rem;
+  text-align: center;
 }
