+++ packages/frontend/src/pages/error/error-info/error-info.component.tsx
@@ -1,13 +1,13 @@
 import { FC } from 'react';
-import { ErrorInfoStatusComponent } from './error-info-status/error-info-status.component';
-import { ErrorInfoTitleComponent } from './error-info-title/error-info-titile.component';
+import { ErrorInfoStatus } from './error-info-status/error-info-status.component';
+import { ErrorInfoTitle } from './error-info-title/error-info-title.component';
 import { IErrorInfo } from '@pages/error/error-info/error-info';
 
-export const ErrorInfoComponent: FC<IErrorInfo> = ({ status = 404,  title}) => {
+export const ErrorInfo: FC<IErrorInfo> = ({ status = 404, title }) => {
   return (
     <div>
-      <ErrorInfoStatusComponent status={status} />
-      <ErrorInfoTitleComponent title={title} />
+      <ErrorInfoStatus status={status} />
+      <ErrorInfoTitle title={title} />
     </div>
   );
 };
