+++ packages/frontend/src/pages/error/error.page.tsx
@@ -1,16 +1,17 @@
 import { isRouteErrorResponse, useRouteError } from 'react-router-dom';
 import css from './error.module.css';
-import { SongTextComponent } from './song-text/song-text.component';
-import { ErrorInfoComponent } from './error-info/error-info.component';
+import { SongText } from './song-text/song-text.component';
+import { ErrorInfo } from './error-info/error-info.component';
+import { FC } from 'react';
 
-export const Error = ({ title = 'Такой страницы не существует.' }) => {
+export const Error: FC = () => {
   const error = useRouteError();
   if (!isRouteErrorResponse(error)) return <h1 className={css.page}>{'Это какая-то странная ошибка. :('}</h1>;
 
-  if (!isRouteErrorResponse(error)) return (
+  return (
     <main className={css.page}>
-      <ErrorInfoComponent status={error.status} title={title} /> */}
-      <SongTextComponent />
+      <ErrorInfo status={error.status} title={error.statusText} />
+      <SongText />
     </main>
   );
 };
