--- packages/frontend/src/pages/splash-screen/splash-screen.d.ts
@@ -1,3 +1,0 @@
-import { ILoader } from '../../components/ui-kit/loader/loader';
-
-export interface ISplashScreen extends ILoader{}
