--- packages/frontend/src/pages/splash-screen/splash-screen.module.css
@@ -1,23 +1,0 @@
-.page {
-  display: flex;
-  height: 100vh;
-  justify-content: center;
-  align-items: center;
-}
-
-.page-text {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 15px;
-  font-size: 2rem;
-}
-
-.page-text__contact {
-  font-size: 1.5rem;
-  color: #444444;
-}
-
-.page-text__link {
-  margin-left: 0.5em;
-}
