+++ packages/frontend/src/pages/splash/splash.d.ts
@@ -1,0 +1,3 @@
+import { ILoader } from '../../components/ui-kit/loader/loader';
+
+export interface ISplashScreen extends ILoader{}
